Squashed commit of the following:
commit 967b9013e209ec96a722cda0d418cb4f0fb76be4
Author: AJ Slater <aj@slater.net>
Date:   Thu Oct 20 23:52:00 2022 -0700

    fix logging deadlock with central logger

commit 7872536293a964465da0ecce1a1bfd04d6b61ae3
Author: AJ Slater <aj@slater.net>
Date:   Thu Oct 20 23:21:24 2022 -0700

    update dev dockerfile
