+++ codex/logger/__init__.py
@@ -1,0 +1,1 @@
+"""Central logger thread."""
