+++ codex/settings/logging.py
@@ -2,67 +2,12 @@
 import logging
 import os
 
-from logging.handlers import RotatingFileHandler
+from logging.handlers import QueueHandler
 
-from colors import color
+from codex.logger.log_queue import LOG_QUEUE, CodexLogger
 
 
-LOG_FMT = "{asctime} {levelname:8} {message}"
-DATEFMT = "%Y-%m-%d %H:%M:%S %Z"
 LOG_EVERY = 5
-LOG_MAX_BYTES = 10 * 1024 * 1024
-VERBOSE: int = int((logging.INFO + logging.DEBUG) / 2)
-
-
-class CodexLogger(logging.Logger):
-    """Custom logger."""
-
-    def verbose(self, message, *args, **kwargs) -> None:
-        """Verbose logging level function."""
-        if self.isEnabledFor(VERBOSE):
-            self._log(VERBOSE, message, args, **kwargs)
-
-
-logging.setLoggerClass(CodexLogger)
-
-
-class ColorFormatter(logging.Formatter):
-    """Logging Formatter to add colors and count warning / errors."""
-
-    FORMAT_COLORS = {
-        "CRITICAL": {"fg": "red", "style": "bold"},
-        "ERROR": {"fg": "red"},
-        "WARNING": {"fg": "yellow"},
-        "INFO": {"fg": "green"},
-        "VERBOSE": {"fg": "cyan"},
-        "DEBUG": {"fg": "black", "style": "bold"},
-        "NOTSET": {"fg": "blue"},
-    }
-    FORMATTERS = {}
-
-    def __init__(self, format, **kwargs):
-        """Set up the FORMATS dict."""
-        super().__init__(**kwargs)
-        for level_name, args in self.FORMAT_COLORS.items():
-            levelno = getattr(logging, level_name)
-            template = color(format, **args)
-            formatter = logging.Formatter(fmt=template, **kwargs)
-            self.FORMATTERS[levelno] = formatter
-
-    def format(self, record):
-        """Format each log message."""
-        formatter = self.FORMATTERS[record.levelno]
-        return formatter.format(record)
-
-
-def _get_file_log_handler(log_dir):
-    """Get the log handlers for initialization."""
-    log_dir.mkdir(exist_ok=True, parents=True)
-    fn = log_dir / "codex.log"
-    file_handler = RotatingFileHandler(fn, maxBytes=LOG_MAX_BYTES, backupCount=30)
-    formatter = logging.Formatter(LOG_FMT, style="{", datefmt=DATEFMT)
-    file_handler.setFormatter(formatter)
-    return file_handler
 
 
 def _get_log_level(debug):
@@ -77,22 +22,11 @@ def _get_log_level(debug):
     return level
 
 
-def _get_log_handlers(log_dir):
-    """Get handlers."""
-    log_file_handler = _get_file_log_handler(log_dir)
-    log_console_handler = logging.StreamHandler()
-    log_formatter = ColorFormatter(LOG_FMT, style="{", datefmt=DATEFMT)
-    log_console_handler.setFormatter(log_formatter)
-    return (log_console_handler, log_file_handler)
-
-
-def init_logging(log_dir, debug):
+def init_logging(debug):
     """Initialize logging."""
-    logging.addLevelName(VERBOSE, "VERBOSE")
-    logging.VERBOSE = VERBOSE  # type: ignore
-    handlers = _get_log_handlers(log_dir)
     level = _get_log_level(debug)
-    logging.basicConfig(level=level, handlers=handlers)
+    handlers = (QueueHandler(LOG_QUEUE),)
+    logging.basicConfig(level=level, format="%(message)s", handlers=handlers)
 
 
 def get_logger(*args, **kwargs) -> CodexLogger:
