+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "codex",
-  "version": "0.14.2-a0",
+  "version": "0.14.2-a1",
   "private": true,
   "description": "ui for codex api",
   "scripts": {
